Add client macros
