+++ .dialyzer_ignore.exs
@@ -1,4 +1,5 @@
 [
+  {"lib/http_ex.ex", :no_return},
   ~r/mock*.ex/,
   ~r/logging*.ex/
 ]
