+++ lib/http_ex/backend/default.ex
@@ -3,6 +3,7 @@ defmodule HTTPEx.Backend.Default do
   The default backend for HTTP calls.
   Uses actual the actual http implementation.
   """
+  import HTTPEx.Clients, only: [def_request: 0]
 
-  use HTTPEx.Backend.Client
+  def_request()
 end
